requirements.txt
streamlit
openai
pandas
ffmpeg-python
imageio-ffmpeg
python-dotenv
whisper @ git+https://github.com/openai/whisper.git